requirements.txt
paramiko~=2.7.2
Jinja2~=2.11.3
netaddr~=0.8.0
PyYAML~=5.4.1